requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==1.4.47
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
